backend/pyproject.toml
[project]
name = "tutan-agent"
version = "0.1.0"
description = "Industrial-grade Android GUI Agent with Semantic Reference System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-socketio>=5.10.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tutan_agent"]
